Opt: 注册时使用截图展示 EULA
修复了一些问题
