+++ src/lang/zh_hans/minigames.yaml
@@ -21,8 +21,14 @@ help:
 # 各小游戏帮助文本
 defuse_tnt:
   description: 拆除 TNT
-  details: （A.K.A wordle）通过猜测排列出正确的密码。
+  details: （wordle）通过猜测排列出正确的密码。
   usage1: defuse-tnt
+
+
+'2048':
+  description: 2048 小游戏
+  details: 数字合成游戏 —— 2048
+  usage1: '2048'
 
 tol:
   description: 关灯挑战
