+++ src/plugins/nonebot_plugin_finding_the_trail/utils/image/generator.py
@@ -19,20 +19,25 @@ from PIL import Image
 from ..enums import Blocks
 from ...config import config
 from nonebot_plugin_localstore import get_config_dir
+from nonebot.log import logger
+import traceback
 
 # 加载默认资源
 base_path = config.ftt_assets_path or get_config_dir("nonebot-plugin-finding-the-trail")
-BLOCKS = {
-    Blocks.NULL: Image.open(base_path.joinpath("stone_bricks.png")),
-    Blocks.WALL: Image.open(base_path.joinpath("bricks.png")),
-    Blocks.START: Image.open(base_path.joinpath("iron_block.png")),
-    Blocks.END: Image.open(base_path.joinpath("diamond_block.png")),
-    Blocks.PISTON: Image.open(base_path.joinpath("piston_top.png")),
-    Blocks.SAND: Image.open(base_path.joinpath("sand.png")),
-    Blocks.COBWEB: Image.open(base_path.joinpath("cobweb.png")),
-    Blocks.PORTAL: Image.open(base_path.joinpath("portal.png")),
-    Blocks.GOLD_PISTON: Image.open(base_path.joinpath("gold_piston_top.png")),
-}
+try:
+    BLOCKS = {
+        Blocks.NULL: Image.open(base_path.joinpath("stone_bricks.png")),
+        Blocks.WALL: Image.open(base_path.joinpath("bricks.png")),
+        Blocks.START: Image.open(base_path.joinpath("iron_block.png")),
+        Blocks.END: Image.open(base_path.joinpath("diamond_block.png")),
+        Blocks.PISTON: Image.open(base_path.joinpath("piston_top.png")),
+        Blocks.SAND: Image.open(base_path.joinpath("sand.png")),
+        Blocks.COBWEB: Image.open(base_path.joinpath("cobweb.png")),
+        Blocks.PORTAL: Image.open(base_path.joinpath("portal.png")),
+        Blocks.GOLD_PISTON: Image.open(base_path.joinpath("gold_piston_top.png")),
+    }
+except FileNotFoundError:
+    logger.error(f"加载贴图失败（文件缺失）: {traceback.format_exc()}")
 
 
 def generate_map_image(game_map: list[list[Blocks]]) -> Image.Image:
