+++ src/plugins/nonebot_plugin_minigames/minigames.json
@@ -8,5 +8,10 @@
     "id": "tol",
     "name": "tol",
     "usage": 1
+  },
+  {
+    "id": "2048",
+    "name": "2048",
+    "usage": 1
   }
 ]
