--- src/plugins/nonebot_plugin_minigames/plugins/nonebot_plugin_2048/help.yaml
@@ -1,3 +1,0 @@
-plugin: 2048
-commands:
-  2048: help;1
