+++ src/plugins/nonebot_plugin_registry/__init__.py
@@ -12,5 +12,7 @@ require("nonebot_plugin_orm")
 require("nonebot_plugin_larklang")
 require("nonebot_plugin_larkutils")
 require("nonebot_plugin_larkuser")
+require("nonebot_plugin_preview")
+require("nonebot_plugin_alconna")
 
 from . import __main__
